Renames DocumentationMerger and moves it to base plugin
