+++ core/src/main/kotlin/transformers/documentation/DocumentableMerger.kt
@@ -3,6 +3,6 @@ package org.jetbrains.dokka.transformers.documentation
 import org.jetbrains.dokka.model.Module
 import org.jetbrains.dokka.plugability.DokkaContext
 
-interface DocumentationNodeMerger {
+interface DocumentableMerger {
     operator fun invoke(modules: Collection<Module>, context: DokkaContext): Module
 }
